Updated with discharge Schottky diodes
Discharge Schottky diodes are needed even if the switching is not done by a MOS-Fet or BJT. The back EMF of the motor seemingly found its way along the power supply and was able to reset the control IC's during the tests. It is recommended to use 1N5817, 1N5818 or 1N5819 depending on the availability. Other more powerful Schottky diodes have not been deemed necessary until now.